# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "title": "http-3rdparty-task-service API",
        "contact": {
            "name": "API Support",
            "email": "zduisekov@gmail.com"
        }
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
        "/accounts/{id}": {
            "get": {
                "description": "get string by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "Show an account",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Account ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/task": {
            "get": {
                "description": "назначение задачи.",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "task"
                ],
                "summary": "Назначить задачу",
                "responses": {
                    "200": {
                        "description": "успешный ответ",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "ошибка запроса",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}